+++ RMSCapstone/resources/views/livewire/admin/roles/view-roles.blade.php
@@ -5,18 +5,20 @@
             <div class=" bg-white-500 relative shadow-md sm:rounded-lg overflow-hidden">
 
                 <!-- Create Role Button -->
-                <div class="flex items-center justify-between p-4">
-                    <button class="bg-blue-600 text-white px-4 py-2 rounded-lg shadow-md hover:bg-blue-700 transition"
-                        onclick="window.location.href='{{ route('admin.create-role') }}'">
-                        + Create Role
-                    </button>
-                </div>
+                @can('role-create')
+                    <div class="flex items-center justify-between p-4">
+                        <button class="bg-blue-600 text-white px-4 py-2 rounded-lg shadow-md hover:bg-blue-700 transition"
+                            onclick="window.location.href='{{ route('admin.create-role') }}'">
+                            + Create Role
+                        </button>
+                    </div>
+                @endcan
 
                 {{-- Display Session Message --}}
                 @if (session('message'))
                     <div x-data="{ show: true }" x-init="setTimeout(() => show = false, 3000)" x-show="show"
                         class="fixed top-4 left-1/2 transform -translate-x-1/2 px-4 py-2 rounded-lg shadow-lg 
-                                                                                                                                            {{ session('alert-type') === 'success' ? 'bg-red-500 text-white' : 'bg-green-500 text-white' }}">
+                                                                                                                                                                                                        {{ session('alert-type') === 'success' ? 'bg-red-500 text-white' : 'bg-green-500 text-white' }}">
                         {{ session('message') }}
                     </div>
                 @endif
@@ -131,19 +133,29 @@
                                 </th>
                                 <td class="px-4 py-3">{{ $role->name }}</td>
                                 <td class="px-4 py-3 flex items-center justify-center space-x-4">
+
                                     <!-- View Icon -->
-                                    <i class="fas fa-eye text-blue-500 cursor-pointer" wire:navigate
-                                        href="{{ route('admin.view-role', ['role' => $role->id]) }}">
-                                    </i>
+                                    @can('role-view')
+                                        <i class="fas fa-eye text-blue-500 cursor-pointer" wire:navigate
+                                            href="{{ route('admin.view-role', ['role' => $role->id]) }}">
+                                        </i>
+                                    @endcan
 
                                     <!-- Edit Icon -->
-                                    <i class="fas fa-edit text-blue-500 cursor-pointer" wire:navigate
-                                        href="{{ route('admin.edit-role', ['role' => $role->id]) }}">
-                                    </i>
+                                    @can('role-edit')
+                                        <i class="fas fa-edit text-blue-500 cursor-pointer" wire:navigate
+                                            href="{{ route('admin.edit-role', ['role' => $role->id]) }}">
+                                        </i>
+                                    @endcan
+
+
                                     <!-- Delete Icon -->
-                                    <i class="fas fa-trash-alt text-red-500 cursor-pointer"
-                                        wire:click="deleteRole('{{ $role->id }}')">
-                                    </i>
+                                    @can('role-delete')
+                                        <i class="fas fa-trash-alt text-red-500 cursor-pointer"
+                                            wire:click="deleteRole('{{ $role->id }}')">
+                                        </i>
+                                    @endcan
+
 
                                 </td>
                             </tr>
